Resolves conflicts from the structural refactor (activitypub/, core/ to
services/*) and the plugin host PR #4912 landing on develop. Featured
Streams tab in Desktop/Mobile content renders before plugin-contributed
tabs. AP Offer (PING) handling and staleness check ported to the new
services/activitypub/ shape. db/query.sql regenerated; new migration
added under persistence/migrations/ for the federated_servers table.

Key resolutions:

services/activitypub/inbox: Offer/Accept/Reject/Leave handlers added as
methods on Service and wired into the resolver.Resolve call in worker.go.

services/activitypub/outbox: follow.go, ping.go, ticker.go ported from
the deleted activitypub/outbox tree; ticker state lives on the Service.

services/activitypub/utils: nodeinfo fetch/validate helpers moved to
the new tree.

services/activitypub/controllers/nodeinfo.go: kept develop's method
receiver convention plus the PR's federation.username field on v2.

persistence/federatedserversrepository: swapped core/data for
services/datastore, made the int32 ID an int64 to match the regenerated
sqlc model, added SetGlobalInstance so inbox handlers and admin helper
can reach the repo without threading deps through every caller. Same
temporary global wired onto configrepository.

services/stalefeaturedcheckservice + activitypub Offer ping: stale check
keeps its package-level Start/Stop. main.go calls them around the
stream lifecycle. activitypub.Service exposes Start/StopStreamPingTicker
so services/stream/stream.go can drive them on connect/disconnect.

persistence/migrations: 00002_featured_streams.sql creates
federated_servers and adds ap_followers.owncast_server. Tests updated
to expect goose v2.

db: hand-merged query.sql across both sides (kept develop's ? bindvars
plus the PR's new federated_servers queries and the owncast_server
column on ap_followers selects/inserts) and regenerated models.go and
query.sql.go via sqlc.

webserver/handlers/admin/federatedservers.go: rewritten as methods on
*Admin using the new outbox method and stream.GetStatus().Online for
the live flag, dropping the core.IsStreamConnected reference.
webserver/handlers/handler.go wires these through s.h.middleware
RequireAdminAuth and s.h.admin.

Removed activitypub/ (leftover PR-side directory) and the misplaced
persistence/legacymigrations/federatedservers.go (federated_servers
moves to a goose migration in the new tree).

lefthook.yml: scoped staged-files golangci-lint to skip the tools/
submodule (its own Go module, breaks the main-module typecheck).

+++ persistence/federatedserversrepository/federatedserversrepository.go
@@ -12,7 +12,7 @@ type FederatedServersRepository interface {
 	GetFederatedServer(iri string) (*models.FederatedServer, error)
 	AddFederatedServer(iri, name, logoURL string, followedAt time.Time, pending bool, username, followStatus string) error
 	UpdateServerStatus(iri string, isOnline bool, metadata *models.FederatedStreamUpdate) error
-	RemoveFederatedServer(id int32) error
+	RemoveFederatedServer(id int64) error
 	RemoveFederatedServerByIRI(iri string) error
 	UpdateFollowStatus(iri, followStatus string, pending bool, acceptedAt, rejectedAt *time.Time) error
 	UpdateServerMetadata(iri, name, displayName, summary, logoURL string) error
